requirements.txt
pymongo
reduct-py==1.12.0